+++ apm-agent-core/src/main/java/co/elastic/apm/agent/impl/transaction/TraceContext.java
@@ -27,10 +27,12 @@ package co.elastic.apm.agent.impl.transaction;
 import co.elastic.apm.agent.impl.ElasticApmTracer;
 import co.elastic.apm.agent.impl.sampling.Sampler;
 import co.elastic.apm.agent.util.HexUtils;
+import com.blogspot.mydailyjava.weaklockfree.WeakConcurrentMap;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 
 import javax.annotation.Nullable;
+import java.lang.ref.WeakReference;
 import java.util.concurrent.Callable;
 
 /**
@@ -56,6 +58,10 @@ public class TraceContext extends TraceContextHolder {
     private static final int PARENT_ID_OFFSET = 36;
     private static final int FLAGS_OFFSET = 53;
     private static final Logger logger = LoggerFactory.getLogger(TraceContext.class);
+    /**
+     * Helps to reduce allocations by caching {@link WeakReference}s to {@link ClassLoader}s
+     */
+    private static final WeakConcurrentMap<ClassLoader, WeakReference<ClassLoader>> classLoaderWeakReferenceCache = new WeakConcurrentMap.WithInlinedExpunction<>();
     private static final ChildContextCreator<TraceContextHolder<?>> FROM_PARENT = new ChildContextCreator<TraceContextHolder<?>>() {
         @Override
         public boolean asChildOf(TraceContext child, TraceContextHolder<?> parent) {
@@ -101,6 +107,9 @@ public class TraceContext extends TraceContextHolder {
     private final StringBuilder outgoingHeader = new StringBuilder(TRACE_PARENT_LENGTH);
     private byte flags;
     private boolean discard;
+    // weakly referencing to avoid CL leaks in case of leaked spans
+    @Nullable
+    private WeakReference<ClassLoader> applicationClassLoader;
 
     /**
      * Avoids clock drifts within a transaction.
@@ -227,6 +236,7 @@ public class TraceContext extends TraceContextHolder {
         id.setToRandomValue();
         clock.init(parent.clock);
         serviceName = parent.serviceName;
+        applicationClassLoader = parent.applicationClassLoader;
         onMutation();
     }
 
@@ -246,6 +256,7 @@ public class TraceContext extends TraceContextHolder {
         discard = false;
         clock.resetState();
         serviceName = null;
+        applicationClassLoader = null;
     }
 
     /**
@@ -362,6 +373,7 @@ public class TraceContext extends TraceContextHolder {
         discard = other.discard;
         clock.init(other.clock);
         serviceName = other.serviceName;
+        applicationClassLoader = other.applicationClassLoader;
         onMutation();
     }
 
@@ -405,6 +417,26 @@ public class TraceContext extends TraceContextHolder {
     @Override
     public Span createSpan(long epochMicros) {
         return tracer.startSpan(fromParent(), this, epochMicros);
+    }
+
+    void setApplicationClassLoader(@Nullable ClassLoader classLoader) {
+        if (classLoader != null) {
+            WeakReference<ClassLoader> local = classLoaderWeakReferenceCache.get(classLoader);
+            if (local == null) {
+                local = new WeakReference<>(classLoader);
+                classLoaderWeakReferenceCache.putIfAbsent(classLoader, local);
+            }
+            applicationClassLoader = local;
+        }
+    }
+
+    @Nullable
+    public ClassLoader getApplicationClassLoader() {
+        if (applicationClassLoader != null) {
+            return applicationClassLoader.get();
+        } else {
+            return null;
+        }
     }
 
     public interface ChildContextCreator<T> {