+++ apm-agent-plugins/apm-jms-plugin/src/test/java/co/elastic/apm/agent/jms/ActiveMqArtemisFacade.java
@@ -152,6 +152,11 @@ public class ActiveMqArtemisFacade implements BrokerFacade {
     }
 
     @Override
+    public boolean shouldTestReceiveNoWait() {
+        return true;
+    }
+
+    @Override
     public Message receiveNoWait(Destination destination) {
         return context.createConsumer(destination).receiveNoWait();
     }